업데이트(화면/기능 추가)

+++ practiceBack/src/main/java/com/example/practiceBack/service/DeviceService.java
@@ -1,9 +1,13 @@
 package com.example.practiceBack.service;
 
+import com.example.practiceBack.dto.Asset;
 import com.example.practiceBack.dto.Device;
+import com.example.practiceBack.mapper.AssetMapper;
 import com.example.practiceBack.mapper.DeviceMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -12,9 +16,11 @@ public class DeviceService {
 
     @Autowired
     private final DeviceMapper deviceMapper;
+    private final AssetMapper assetMapper;
 
-    public DeviceService(DeviceMapper deviceMapper) {
+    public DeviceService(DeviceMapper deviceMapper, AssetMapper assetMapper) {
         this.deviceMapper = deviceMapper;
+        this.assetMapper = assetMapper;
     }
 
     // 장비 DB 전부 검색
@@ -29,6 +35,10 @@ public class DeviceService {
 
     // 장비 DB 삭제
     public void deleteDeviceBySerial(String serialNumber) {
+        String assetSn = assetMapper.findAssetSnByDeviceSn(serialNumber);
+        if (!assetSn.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, assetSn);
+        }
         deviceMapper.deleteBySerialNumber(serialNumber);
     }
 }
